MXE-20 Made JSONAuditLog::append safer
Changed to a mutex to protect against a race condition (between len() and
write()), and implemented error handling and retry on EAGAIN, EWOULDBLOCK,
and EINTR.

We currently blow up the server if we can't write to the audit log.  To
not do this would allow authorization failures to go unnoticed, and would
allow a server with e.g. a full audit log disk to repeatedly return errors
to legitimate clients on DDL operations.  By shutting down, we allow
failover to take control and handle the situation nicely.